A face generation, which can change the appearance of an existing person through face generation.
It can be used for plastic surgery, and some appearance analysis is mainly used for criminals, and infers the suspect of fugitive by recognizing his son’s face. people. It can also be used for abducting and trafficking children in the early years. Through the appearance of the parents of the child, the app can be used to infer the current appearance of the child and be identified from big data.